assignment10: Main: v0.1d
Re-work what the program does with input. I think it’s better that the
input is read in once, and then passed to all the constructors, rather
than having a method call in each constructor to ask for input.

However, Utility.getInput() is not set up to be called multiple times.
The BufferedReader is closed at the end of the method call, and we
don’t create a new one each time. As a result, if for some reason we
didn’t pass the input into a constructor after having already called
Utility.getInput(), the program will throw an error.